dog-detection/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "dog-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "gcloud-aio-storage>=9.6.4",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.21",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 95
fix = true
exclude = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
docstring-code-format = true


